+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/fargate/validate_04_estimates.py
@@ -68,7 +68,7 @@ EXPECTED_SCHEMA = pb.Schema(
         ),
         IndCqcColumns.id_per_locationid_import_date: "UInt32",
         IndCqcColumns.main_job_role_clean_labelled: str(
-            CategoricalColumnTypes.JobRoleEnumType
+            CategoricalColumnTypes.JobRoleCatType
         ),
         IndCqcColumns.ascwds_job_role_counts: "Int16",
         IndCqcColumns.job_role_filtering_rule: str(
@@ -87,7 +87,7 @@ EXPECTED_SCHEMA = pb.Schema(
         IndCqcColumns.estimate_filled_posts_from_all_job_roles: "Float64",
         IndCqcColumns.difference_estimate_filled_posts_and_from_all_job_roles: "Float64",
         IndCqcColumns.main_job_group_labelled: str(
-            CategoricalColumnTypes.JobGroupEnumType
+            CategoricalColumnTypes.JobGroupCatType
         ),
         PartitionKeys.year: "Int64",
     }
@@ -308,10 +308,10 @@ def other_validation(
         )
         .col_vals_between(
             columns=IndCqcColumns.difference_estimate_filled_posts_and_from_all_job_roles,
-            left=-0.0002,
+            left=-0.002,
             right=1,
             na_pass=True,
-            brief="Difference between estimate_filled_posts and estimate_filled_posts_from_all_job_roles should be between -0.0002 and 1 where present",
+            brief="Difference between estimate_filled_posts and estimate_filled_posts_from_all_job_roles should be between -0.002 and 1 where present",
         )
         # Date plausibility
         .col_vals_ge(